refactor(engine): move the OnIncoming tests and listDir out of cleanup_test.go
cleanup_test.go's header says "partial-file cleanup tests", and eight of its
ten tests are that. The other two pin the OnIncoming contract (it fires
exactly once, with the batch summary, before the first OnProgress call) and
say nothing about what is left on disk; they and their only helper, countOf,
now live in incoming_test.go. listDir is used by cleanup_test.go,
hostile_test.go and partial_test.go alike, so it joins newConnectedPair and
runTransfer in loopback_test.go, the file the other test files already point
at for the harness. The new file carries no GOOS token, and its note sits
below the package clause.

Pure move: the sorted go test -list output is identical on Windows and
GOOS=linux, the -v PASS list is identical, go doc -all -u is unchanged, and
git diff --color-moved shows only package, import and header lines as new.
